feat(tabs): add custom styled variant
